Use child_ops data to order children in diffs
Assuming we find the correct data in the diff structure, we can properly order
the children of nodes in the diff section. Previously, we attempted to derive
a correct order by looking only at the labels. We'll be able to remove this
logic once all diff data is updated to contain the child_ops data.