Begin to modularize layer data by adding a setting
Previously, the mega `generator` module knew about most of the different types
of layers. It created a big dictionary of them for iterating through. This
patch moves that listing into a settings file so that it can be replaced by
different clients. Further, it moves data related to where the layer
information comes from and how it should be applied into the individual layer
classes; it also introduces a few constants and creates a common abstract base
class for most layers.

This does _not_ make the section-by-section analyses on the right hand side
configurable; they are currently hard coded into the sidebar view.