Define behavior when nodes are re-ordered
Previously, when a node was re-ordered within its parent, the UI would see the
same node as being present in two locations (where it had been deleted from
and where it was added to). This would silent fail the diff ordering process,
which would then re-order all of the children using a best-guess heuristic.

This patch defines behavior for such a change, placing the node in its final
position when displaying the diff.