Use CSS for paragraph markers
We have been relying on the markup to tell the browser which marker type to
use. Aside from an inline style, this worked fine, but we're reaching
situations where a containing div needs the ability to override the markers of
its contents. Enter CSS.

Also swapped an `em` for `span` when rendering the paragraph marker. This
element hasn't ever been displayed (though we'll start soon); it shouldn't
have ever been an `em`

Also removes an `extra` attribute on `ol`s. This shouldn't have ever been
present -- it's unused and renders unnecessary content.