Explicitly tag collapsed paragraphs
We've relied on some intelligent style sheet magic to implement collapsed
markers (where a paragraph and its subparagraph are displayed on the same
line, visually) but that only works when using the list-style marker generated
by the browser. As we'll need to use markers from the text at multiple points
in the near future, this patch explicitly tags collapsed paragraphs when
rendering the markup.

This should have no affect on existing templates.