Allow node-type specific templates
Previously, we added functionality for node-specific templates which could be
overridden by a client. This patch updates that functionality to allow
node-type specific templates, i.e. selecting the node rendering template
automatically based on the node type. It moves around multiple templates to
make this easier -- we look for a templates in the `regulations/tree/` dir.