Duplicate notice compilation here
Looks like there's a shared code snippet between the notice compilation in
parser and this tree builder for diffs. As there's a change in the notice
compilation, we need to also change it here. This logic should be pulled out
into a shared repository or data structure, but that's outside the scope of
the current issue