Allow sidebar view to be configured
Previously, we have only allowed a single, site-wide configuration for the
sidebar. As we really want to have slightly different sidebars for different
sections of the site (regs vs diffs vs search results, etc.), we'd like to add
the ability to configure the SideBarView on instantiation.

This also removes the "custom-sidebar" (which is used by CFPB but could be
deleted) and allows the sidebar components to be defined as _either_ a class
path _or_ a full class object