Reduce special cases in sidebar JS
Previously, a sidebar sub-view needed to exist for every type of sidebar
component, whether that component needed special functionality or not. This
removes that requirement; sxs and definitions are still special, but help,
rulings, and potentially others aren't.

To accomplish that goal, this patch removes logic from the JS so that it is
works with a largely opaque sidebar <section> (with a unique identifier for
cache purposes) rather than having deep knowledge of the each component within
that <section>. As a bonus, this should also aid the efforts to make the
landing page sidebars return.