Fix issue w/ sidebar init; simplify it
We had been calling `openRegFolders` on Sidebar init. This function was also
used when AJAX-ing in a new sidebar, meaning that it would remove the existing
sidebar components before adding any "new" components. After a bit of poking
around, I'm still not sure why this wasn't a problem before (when all of the
components were Backbone views), but I'd rather just fix it here.

This updates the init to _not_ pass through the `openRegFolders` function
(which didn't do any init other than creating a SxSList view). It also renames
`closeChildren` to `removeChildren` to make it more clear what that function
does.