Return version info until the last moment
Version info has been more or less removed from the layer api -- it becomes
part of the opaque document identifier. The first pass for integrating that
into -site was to use the same opaque identifier throughout. That doesn't work
well, however, as version information is very important to -site's
functioning. This resulted in lots of string joins and splits.

This patch replaces the previous logic to carry version information throughout
the app, only converting it into the identifier used by the API immediately
before sending a request.