Replace --dev flag with settings.DEBUG_JS.
We introduced a `--dev` flag to `compile_frontend` in 2bcb424, but this
flag should be coupled to the existing `DEBUG_JS` settings variable.
This patch reverts the `--dev` flag and uses `DEBUG_JS` in its place.
The `--no-install` flag is retained, since there isn't a corresponding
settings variable.

h/t @cmc333333 @tadgh-ohiggins @xtine