Add CFR changes table of contents
Very rough pass at adding a list of modified sections/authorities to the
drawer when viewing a proposed rule. This ToC is built from the list of
changes, the meta data, and the toc data for the "new" version of the
regulation.

Pain points which this punts on:
1) "timeline" is the wrong name for this tab, but it's one that the JS
understands. Rather than try to figure out how to generalize that, this just
co-opts the term.
2) The order of "changes" matters -- we previously used a dictionary. This
assumes a list of key-value pairs is used instead and converts any
dictionaries into that format