Split paragraph marker layer into two
Previously, we had a paragraph marker layer which would convert "(a)" into:
    <p>
        <span class="stripped-marker">a.</span>
        <span class="paragraph-marker">(a)</span>
        Paragraph text here...
    </p>

We'd then hide the latter and float out the former. With text that collapses
to highlight changes, we can't use this strategy too easily; we often need to
show _only_ the marker and none of the text. To get there, this patch splits
the paragraph marker layer into two steps. The first hides the original
marker, while the second makes the marker meta-data of the node. The end
result looks like:
    <p>
        <span class="stripped-marker">a.</span>
        <span class="paragraph-text">
            <span class="paragraph-marker">(a)</span>
            Paragraph text here...
        </span>
    </p>

We keep the old layer for backwards compatibility, but can remove it once
we've removed references in ATF/FEC.