Fetch field options from regs.gov.
This patch fetches options and dependency-contingent options for comment
fields from the regs.gov api. Field options are added to the template
context so that template overrides can access them if needed. Note: as
discussed with @cmc333333 and @vrajmohan, it would make sense to fetch
field metadata once in -parser instead; I'm proposing this change in the
interest of getting something working quickly.