Pass empty strings to template for missing values.
Django renders `None` as `"None"` unless we use the `default` or
`default_if_none` filters. This leads to an error when comment
submission fails such that the metadata url is empty but appears not to
be, since it shows up as `"None"`. This patch uses empty strings instead
of `None` when template variables are empty.