Tweak ProseMirror menu
Makes ProseMirror's menu hang out at the top of the text box rather than be
inline. This also adds a change to the pdf template (so that it uses the same
class for comment content) and to CSS styles around li/ols (which are
list-style-type: none elsewhere in the app).