Rework "dependency" logic on N&C submission
Previously, we had shown/hidden the select and the text input associated with
a "dependent" individually. We need to be able to show divs _containing_ these
fields now. This patch modifies the logic to be less input-specific; it also
modifies the logic around the text input (dubbed "writeIn") so that it points
to a specific select.