Align -site with -parser's notice output
As we were using mock data to represent notices for several weeks, field names
were a bit out of sync. Now that we have real data, make sure we're using the
proper fields.

This also constructs a `cfr_refs` field. We probably want to add such a
structure to the parsed notices at some point, but we can punt for a bit.