Switch meta.accepts_comments into 4 states
This uses the `enum` model introduced in Python 3.4 (but since backported).
I'm definitely not wedded to this, but I think it's a good model.

Mostly template changes. I'm not a fan of using `.name == 'MyValue`, but
Django's templates will silently fail if we were to do something like
meta.comment_state == CommentState.Mispelled
so I don't think we're actually losing anything with the string comparison.

Removes the "regulations/comment-write-closed.html" partial; we're preferring
blocks moving forward. See
https://github.com/eregs/regulations-site/issues/358

Tests need to be updated