Add template for repeated submission failure
If we try to submit to regs.gov multiple times, we eventually give up. The
worker process will notify the user by way of S3 that such an event has
occurred.

Adds a third parameter to the template replacement code if we want to target a
specific template (useful when there are nested divs).