Add warnings for non-JS, non-localstorage users
We don't currently support this audience, so we should warn them where
relevant. For the latter, we could build another custom version of Modernizer,
but I'm not seeing any of our existing JS code check it, so I'm guessing it's
being phased out. Easy enough to write a manual test.

To test locally, try Safari's Privacy mode and write a comment.