Add "daily" cache, expiring at midnight or sooner
We'd like to use the "long term" cache for notice-and-comment data, as
rendering content often takes a significant amount of time. Unfortunately, the
content of these pages may change at midnight (due to the close of comment
period, for example), so we need to extend the existing cache middleware to
take into account the current date.