Move "simple" mixins into base-variables
We have a few font-related mixins defining sans and serif fonts used in the
app. We then build new mixins on top of them. Unfortunately, that means that,
if we want to define a different font, we need to define not only the base
mixins but also all of the mixins built on top.

This is error prone and not sustainable. To resolve, this patch moves those
font mixins into the "base-variables" file, to be overridden in "variables".
This allows clients (such as ATF) to redefine *just* the fonts and have
everything build on top of that.