Migrate definition popup to match sidebar
For context, we migrated most of the sidebar to use a different style of
templates many moons ago. The idea was to make those sidebars data-driven
rather than having custom JS for each. In any event, we didn't migrate the
definition widget, but that time has now come.

Remove all of the special classes related to the definitions, including the
`continue-link` and `sidebar-header`. Use a cf-icon rather than an image for
the close button.