Return color to rhs definition while loading
The markup and CSS rules around the rhs definition loading were standardized
a bit in #461. Those tweaks didn't account for the short-lived header that
appears while loading the definitions (nor the header that appears when
there's a network error). This modifies the stylesheet to apply to those
situations as well. It also makes the JS tags match the later markup