Begin to replace LayerCreators
These classes don't provide a good interface to working with layers of data,
so they're on the chopping block. As a first step, however, we need to put a
shim in between them and the functions that call them. This will allow us to
follow up and remove the classes altogether.