Standardize layer methods
Previously, we had a relatively convoluted system where layers were grouped
into one of three categories: inline, search-replace, and paragraph. We had a
"LayerApplier" for each style, which knew which methods to call on its style
of Layer. The HTMLBuilder knew about the three LayerAppliers and called each
separately.

This begins to replace all of this by standardizing Layer objects to have two
methods:
* inline_replacements: to be used when replacing chunks of the paragraph text
* attach_metadata: to be used when attaching metadata to the whole node (e.g.
  for TOC)

There's still a lot of cleanup to do, such as removing the LayerApplier
classes and modifying the DiffLayers to match, but this is a good checkpoint.