Add grunt-cli as a dependency.
Rather than requiring users install grunt-cli globally (which can introduce
version problems), we'll be explicit that it's a dependency. Since grunt may
not be in the global env anymore, we'll also switch to using package.json's
`scripts` attr.

This bumps out version to 8.4.2 in preparation of our next release.

It resolves #508